Added data text files, added dto classes for data files and changed Database and DatabasePopulateService classes

+++ sql/init_db.sql
@@ -3,7 +3,7 @@ CREATE TABLE worker(
 	name VARCHAR NOT NULL,
 	birthday DATE,
 	level VARCHAR NOT NULL,
-	salary INTEGER
+	salary BIGINT
 );
 ALTER TABLE worker ADD CONSTRAINT check_worker_name CHECK(LENGTH(name) >= 2 AND LENGTH(name) <= 1000);
 ALTER TABLE worker ADD CONSTRAINT check_worker_birthday CHECK(year(birthday) > 1900 );
